Move remaining pipeline steps into pipeline directory
We no longer need the Claude namespace as the question router and
structured answer composer steps now only use Claude models, so we can
move them into the main pipeline directory.
